requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
mysql-connector-python==8.0.26
azure-cosmos==4.5.1